platform/pico2w: address PR #102 review — cleanup, module splits, dep fixes
Quick wins:
- Remove SAVE_SYNC_PLAN.md / WIFI_PLAN.md from source control; gitignore *_PLAN.md
- Drop unused heapless `mpmc_large` feature (migrated to spsc::Queue)
- Remove bug-#5 diagnostic leftovers (DmaState repr/align, copy_dma_step
  #[track_caller]); condense stale §G10 comment; document WATCHDOG_WINDOW_MS
  and COMMAND_QUEUE_CAPACITY rationale

Refactors:
- Extract mpu module (setup_core0_data_mpu + setup_core1_mpu)
- Extract integrity module (boot-time .data CRC guard; IMAGE_CRC symbol preserved)
- Split crash/mod.rs (1151 lines) into record/sector/context submodules
- Deduplicate crash handler (shared commit_crash_and_reset helper)
- Decompose wifi_codec url_decode_into into helpers
- Extract host-testable dhcp module out of wifi_codec
- Replace hand-rolled BusEventQueue with heapless::Vec

Critical section:
- Revert hand-rolled RpSpinlockCriticalSection (verbatim duplicate of embassy's)
  back to embassy-rp's critical-section-impl feature

Verified: core tests 649, pico2w --lib 191, firmware release build clean.
