+++ .gitignore
@@ -1,3 +1,4 @@
 target/
 lcov.info
 docs/dkl2-sprite-investigation.md
+*_PLAN.md
