+++ platform/pico2w/src/crash/mod.rs
@@ -26,13 +26,17 @@ pub mod storage;
 #[cfg(target_arch = "arm")]
 pub mod handler;
 
+pub mod context;
+pub mod record;
+pub mod sector;
+
 // ---------------------------------------------------------------------------
 // Build-time constants (git hash, firmware version).
 // ---------------------------------------------------------------------------
 include!(concat!(env!("OUT_DIR"), "/crash_build_info.rs"));
 
 // ---------------------------------------------------------------------------
-// Protocol constants.
+// Protocol constants (shared across sub-modules).
 // ---------------------------------------------------------------------------
 
 /// Sentinel written to `WATCHDOG.scratch0` by the fault handler.
@@ -48,621 +52,8 @@ pub const SECTOR_MAGIC: [u8; 4] = *b"RCLG";
 /// Fixed size of one crash record in bytes.
 pub const RECORD_SIZE: usize = 128;
 
-/// Fixed size of the sector header in bytes (occupies slot 0).
-pub const SECTOR_HEADER_SIZE: usize = RECORD_SIZE;
-
-/// Number of crash records that fit in one erase sector after the header.
-pub const MAX_RECORDS_PER_SECTOR: usize = 31; // (4096 / 128) - 1
-
 // ---------------------------------------------------------------------------
-// Crash kind and flags.
-// ---------------------------------------------------------------------------
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[repr(u8)]
-pub enum CrashKind {
-    HardFault = 0,
-    Panic = 1,
-    /// The hardware watchdog timer expired before the firmware fed it.
-    /// No ARM exception frame is available — the CPU was reset by hardware,
-    /// not by the fault handler.
-    WatchdogTimeout = 2,
-    /// Boot-time reset cause captured from reset-status registers.
-    ResetReason = 3,
-    /// Core-0 multicore transport pointer triplet was overwritten.
-    TransportSmash = 4,
-    Unknown = 0xFF,
-}
-
-impl CrashKind {
-    pub fn from_u8(v: u8) -> Self {
-        match v {
-            0 => Self::HardFault,
-            1 => Self::Panic,
-            2 => Self::WatchdogTimeout,
-            3 => Self::ResetReason,
-            4 => Self::TransportSmash,
-            _ => Self::Unknown,
-        }
-    }
-}
-
-/// Bit flags describing which fields in a crash record are populated.
-pub mod flags {
-    /// ARM exception frame (pc, lr, cfsr, …) is valid.
-    pub const HAS_ARM_REGS: u8 = 0x01;
-    /// GB CPU registers and cycle counter are valid.
-    pub const HAS_GB_STATE: u8 = 0x02;
-    /// ROM identity (id prefix, bank) is valid.
-    pub const HAS_ROM_INFO: u8 = 0x04;
-    /// Panic file / line is valid.
-    pub const HAS_PANIC_LOC: u8 = 0x08;
-    /// The faulting SP was below its stack limit at crash time — stack overflow.
-    /// When set, `stack_headroom` holds the overflow depth in bytes.
-    /// When clear, `stack_headroom` holds the remaining headroom in bytes.
-    /// The relevant stack (and limit) is core 0's unless [`FAULT_ON_CORE1`] is set.
-    pub const HAS_STACK_OVERFLOW: u8 = 0x10;
-    /// The crash occurred on core 1 (the emulator worker). When set, the stack
-    /// measurement (`stack_headroom` / [`HAS_STACK_OVERFLOW`]) is relative to
-    /// core 1's dedicated stack limit rather than core 0's `_stack_end`.
-    pub const FAULT_ON_CORE1: u8 = 0x20;
-    /// HardFault/DebugMonitor diagnostic tail is present in `panic_loc`:
-    /// `[68..72] = sp_before (faulted-thread SP; corrupted LR slot = sp_before-4)`,
-    /// `[72..76] = stacked r12`.
-    pub const HAS_HARDFAULT_EXTENDED_REGS: u8 = 0x40;
-    /// Stack-protector panic: `arm_lr` contains the return address captured by
-    /// `__stack_chk_fail`, identifying the guarded function whose canary failed.
-    pub const HAS_STACK_CHK_FAIL_LR: u8 = 0x80;
-}
-
-// ---------------------------------------------------------------------------
-// CrashRecord — 128-byte on-flash crash record.
-// ---------------------------------------------------------------------------
-//
-// Byte layout (all multi-byte fields little-endian):
-//
-//   [0..4]    magic              b"RCRP"
-//   [4]       schema_ver         = 1
-//   [5]       crash_kind         CrashKind as u8
-//   [6]       flags              flags::* bitmask
-//   [7]       slot_seq           slot index within this erase cycle (0-30)
-//   [8..11]   fw_version         [major, minor, patch]
-//   [11]      _pad0
-//   [12..16]  git_hash           first 4 bytes of git SHA as u32 LE
-//   [16..20]  arm_pc             program counter at fault
-//   [20..24]  arm_lr             link register at fault
-//                                or stack-canary failure LR when
-//                                HAS_STACK_CHK_FAIL_LR is set
-//   [24..28]  arm_cfsr           Configurable Fault Status Register
-//   [28..32]  arm_hfsr           HardFault Status Register
-//   [32..36]  arm_fault_addr     MMFAR or BFAR (whichever is valid)
-//   [36..40]  _pad1
-//   [40..44]  rom_id_prefix      first 4 bytes of ROM SHA-256
-//   [44..46]  rom_bank           currently mapped ROM bank
-//   [46]      ram_bank
-//   [47]      _pad2
-//   [48]      gb_a               GB accumulator
-//   [49]      gb_f               GB flags byte
-//   [50]      gb_b
-//   [51]      gb_c
-//   [52]      gb_d
-//   [53]      gb_e
-//   [54]      gb_h
-//   [55]      gb_l
-//   [56..58]  gb_sp              GB stack pointer
-//   [58..60]  gb_pc              GB program counter (where in the ROM)
-//   [60..64]  gb_cycle_lo        lower 32 bits of cycle counter
-//   [64]      ppu_ly             current scanline
-//   [65]      ppu_lcdc           LCD control register (0 if unavailable)
-//   [66]      ppu_stat           LCD status register
-//   [67]      _pad3
-//   [68..80]  panic_loc          null-terminated filename (last segment, ≤11 chars)
-//                                or HardFault/DebugMonitor diagnostic words when
-//                                HAS_HARDFAULT_EXTENDED_REGS is set:
-//                                [68..72] pre-handler r4, [72..76] stacked r12
-//   [80..82]  panic_line         source line number
-//   [82..84]  stack_headroom     bytes between MSP-at-fault and _stack_end:
-//                                if HAS_STACK_OVERFLOW set → overflow depth,
-//                                else → remaining stack headroom
-//   [84..88]  dma_busy_mask      bitmask of DMA channels with BUSY=1 at crash time
-//   [88..116] dma_write_addrs    WRITE_ADDR for DMA channels 0-6 (7 × u32 LE)
-//   [116..120] _reserved         zero-filled (schema v2+; 0 in v1 records)
-//   [120..124] crc32             CRC32 of bytes [0..120]
-//   [124..128] _pad4
-//
-// Total: 128 bytes.
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub struct CrashRecord {
-    pub schema_ver: u8,
-    pub crash_kind: u8,
-    pub flags: u8,
-    pub slot_seq: u8,
-    pub fw_version: [u8; 3],
-    pub git_hash: u32,
-    pub arm_pc: u32,
-    pub arm_lr: u32,
-    pub arm_cfsr: u32,
-    pub arm_hfsr: u32,
-    pub arm_fault_addr: u32,
-    pub rom_id_prefix: [u8; 4],
-    pub rom_bank: u16,
-    pub ram_bank: u8,
-    pub gb_a: u8,
-    pub gb_f: u8,
-    pub gb_b: u8,
-    pub gb_c: u8,
-    pub gb_d: u8,
-    pub gb_e: u8,
-    pub gb_h: u8,
-    pub gb_l: u8,
-    pub gb_sp: u16,
-    pub gb_pc: u16,
-    pub gb_cycle_lo: u32,
-    pub ppu_ly: u8,
-    pub ppu_lcdc: u8,
-    pub ppu_stat: u8,
-    /// Last path segment of the panic source file, null-terminated, ≤ 11 chars.
-    pub panic_loc: [u8; 12],
-    pub panic_line: u16,
-    /// Bytes between MSP-at-fault and `_stack_end`.
-    /// Meaning depends on `flags::HAS_STACK_OVERFLOW`:
-    ///   set   → overflow depth (MSP was this many bytes below `_stack_end`)
-    ///   clear → remaining headroom (MSP was this many bytes above `_stack_end`)
-    pub stack_headroom: u16,
-    /// Bitmask of DMA channels (0-7) that had CTRL_TRIG.BUSY=1 at crash time.
-    /// Zero if DMA snapshot was not captured (schema v1 records).
-    pub dma_busy_mask: u32,
-    /// WRITE_ADDR register snapshot for DMA channels 0-6 at crash time.
-    /// Zeros for schema v1 records or channels that were not snapshotted.
-    pub dma_write_addrs: [u32; 7],
-}
-
-/// Opaque 128-byte wire representation.
-pub type RecordBytes = [u8; RECORD_SIZE];
-
-impl CrashRecord {
-    /// Serialise to the 128-byte on-flash layout, computing and embedding CRC32.
-    pub fn to_bytes(&self) -> RecordBytes {
-        let mut buf = [0u8; RECORD_SIZE];
-        buf[0..4].copy_from_slice(&RECORD_MAGIC);
-        buf[4] = self.schema_ver;
-        buf[5] = self.crash_kind;
-        buf[6] = self.flags;
-        buf[7] = self.slot_seq;
-        buf[8..11].copy_from_slice(&self.fw_version);
-        // buf[11] = 0 (pad)
-        buf[12..16].copy_from_slice(&self.git_hash.to_le_bytes());
-        buf[16..20].copy_from_slice(&self.arm_pc.to_le_bytes());
-        buf[20..24].copy_from_slice(&self.arm_lr.to_le_bytes());
-        buf[24..28].copy_from_slice(&self.arm_cfsr.to_le_bytes());
-        buf[28..32].copy_from_slice(&self.arm_hfsr.to_le_bytes());
-        buf[32..36].copy_from_slice(&self.arm_fault_addr.to_le_bytes());
-        // buf[36..40] = 0 (pad)
-        buf[40..44].copy_from_slice(&self.rom_id_prefix);
-        buf[44..46].copy_from_slice(&self.rom_bank.to_le_bytes());
-        buf[46] = self.ram_bank;
-        // buf[47] = 0 (pad)
-        buf[48] = self.gb_a;
-        buf[49] = self.gb_f;
-        buf[50] = self.gb_b;
-        buf[51] = self.gb_c;
-        buf[52] = self.gb_d;
-        buf[53] = self.gb_e;
-        buf[54] = self.gb_h;
-        buf[55] = self.gb_l;
-        buf[56..58].copy_from_slice(&self.gb_sp.to_le_bytes());
-        buf[58..60].copy_from_slice(&self.gb_pc.to_le_bytes());
-        buf[60..64].copy_from_slice(&self.gb_cycle_lo.to_le_bytes());
-        buf[64] = self.ppu_ly;
-        buf[65] = self.ppu_lcdc;
-        buf[66] = self.ppu_stat;
-        // buf[67] = 0 (pad)
-        buf[68..80].copy_from_slice(&self.panic_loc);
-        buf[80..82].copy_from_slice(&self.panic_line.to_le_bytes());
-        buf[82..84].copy_from_slice(&self.stack_headroom.to_le_bytes());
-        buf[84..88].copy_from_slice(&self.dma_busy_mask.to_le_bytes());
-        for (i, &addr) in self.dma_write_addrs.iter().enumerate() {
-            let off = 88 + i * 4;
-            buf[off..off + 4].copy_from_slice(&addr.to_le_bytes());
-        }
-        // buf[116..120] = 0 (reserved)
-        // buf[124..128] = 0 (pad)
-
-        let checksum = crc32(&buf[..120]);
-        buf[120..124].copy_from_slice(&checksum.to_le_bytes());
-        buf
-    }
-
-    /// Deserialise from a 128-byte buffer.
-    ///
-    /// Returns `Err` if the magic is wrong or the CRC32 does not match.
-    pub fn from_bytes(buf: &RecordBytes) -> Result<Self, CrashDecodeError> {
-        if buf[0..4] != RECORD_MAGIC {
-            return Err(CrashDecodeError::BadMagic);
-        }
-        let stored_crc = u32::from_le_bytes(buf[120..124].try_into().unwrap());
-        let computed_crc = crc32(&buf[..120]);
-        if stored_crc != computed_crc {
-            return Err(CrashDecodeError::CrcMismatch {
-                stored: stored_crc,
-                computed: computed_crc,
-            });
-        }
-        Ok(Self {
-            schema_ver: buf[4],
-            crash_kind: buf[5],
-            flags: buf[6],
-            slot_seq: buf[7],
-            fw_version: buf[8..11].try_into().unwrap(),
-            git_hash: u32::from_le_bytes(buf[12..16].try_into().unwrap()),
-            arm_pc: u32::from_le_bytes(buf[16..20].try_into().unwrap()),
-            arm_lr: u32::from_le_bytes(buf[20..24].try_into().unwrap()),
-            arm_cfsr: u32::from_le_bytes(buf[24..28].try_into().unwrap()),
-            arm_hfsr: u32::from_le_bytes(buf[28..32].try_into().unwrap()),
-            arm_fault_addr: u32::from_le_bytes(buf[32..36].try_into().unwrap()),
-            rom_id_prefix: buf[40..44].try_into().unwrap(),
-            rom_bank: u16::from_le_bytes(buf[44..46].try_into().unwrap()),
-            ram_bank: buf[46],
-            gb_a: buf[48],
-            gb_f: buf[49],
-            gb_b: buf[50],
-            gb_c: buf[51],
-            gb_d: buf[52],
-            gb_e: buf[53],
-            gb_h: buf[54],
-            gb_l: buf[55],
-            gb_sp: u16::from_le_bytes(buf[56..58].try_into().unwrap()),
-            gb_pc: u16::from_le_bytes(buf[58..60].try_into().unwrap()),
-            gb_cycle_lo: u32::from_le_bytes(buf[60..64].try_into().unwrap()),
-            ppu_ly: buf[64],
-            ppu_lcdc: buf[65],
-            ppu_stat: buf[66],
-            panic_loc: buf[68..80].try_into().unwrap(),
-            panic_line: u16::from_le_bytes(buf[80..82].try_into().unwrap()),
-            stack_headroom: u16::from_le_bytes(buf[82..84].try_into().unwrap()),
-            dma_busy_mask: u32::from_le_bytes(buf[84..88].try_into().unwrap()),
-            dma_write_addrs: {
-                let mut addrs = [0u32; 7];
-                for (i, addr) in addrs.iter_mut().enumerate() {
-                    let off = 88 + i * 4;
-                    *addr = u32::from_le_bytes(buf[off..off + 4].try_into().unwrap());
-                }
-                addrs
-            },
-        })
-    }
-
-    /// Human-readable crash kind.
-    pub fn crash_kind_name(&self) -> &'static str {
-        match CrashKind::from_u8(self.crash_kind) {
-            CrashKind::HardFault => "HardFault",
-            CrashKind::Panic => "Panic",
-            CrashKind::WatchdogTimeout => "WatchdogTimeout",
-            CrashKind::ResetReason => "ResetReason",
-            CrashKind::TransportSmash => "TransportSmash",
-            CrashKind::Unknown => "Unknown",
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum CrashDecodeError {
-    BadMagic,
-    CrcMismatch { stored: u32, computed: u32 },
-}
-
-// ---------------------------------------------------------------------------
-// SectorHeader — 128-byte header at byte 0 of the crash log flash sector.
-// ---------------------------------------------------------------------------
-//
-//   [0..4]   magic        b"RCLG"
-//   [4..8]   erase_count  how many times this sector has been erased
-//   [8]      next_slot    next free record slot (0-30); 0xFF = full
-//   [9..128] _reserved    zeros
-
-pub const SECTOR_FULL: u8 = 0xFF;
-
-// ---------------------------------------------------------------------------
-// Slot discovery (pure, testable without hardware)
-// ---------------------------------------------------------------------------
-
-/// Return the index of the first slot that is completely erased (all bytes
-/// 0xFF), or `None` if no such slot exists (sector is full or entirely corrupt).
-///
-/// **Why 0xFF instead of "not RCRP"?**
-/// NOR flash can only transition bits from 1 → 0, never 0 → 1 without an
-/// erase.  A slot that contains any non-0xFF byte is considered occupied —
-/// whether it holds a valid RCRP record, a corrupt/partially-written record,
-/// or garbage from a failed previous write.  Selecting such a slot as the
-/// write target would AND the new data with the residual bits, silently
-/// corrupting the magic and CRC without returning any error.
-///
-/// When all 31 slots are occupied (by either valid records or non-erased
-/// corruption), this returns `None`, triggering a sector erase in the caller
-/// before a fresh record is written.
-///
-/// The caller reads the leading 4 bytes of each record slot from flash and
-/// passes them here.  Keeping this logic pure (no I/O) lets it be unit-tested
-/// on the host without any embedded hardware or flash driver.
-pub fn find_next_empty_slot(slot_magics: &[[u8; 4]; MAX_RECORDS_PER_SECTOR]) -> Option<usize> {
-    slot_magics
-        .iter()
-        .position(|m| m == &[0xFF, 0xFF, 0xFF, 0xFF])
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct SectorHeader {
-    pub erase_count: u32,
-    /// Next free record slot index (0-30).  `SECTOR_FULL` when no room remains.
-    pub next_slot: u8,
-}
-
-impl SectorHeader {
-    /// Fresh header for a newly erased sector.
-    pub fn fresh(erase_count: u32) -> Self {
-        Self {
-            erase_count,
-            next_slot: 0,
-        }
-    }
-
-    pub fn to_bytes(&self) -> RecordBytes {
-        let mut buf = [0u8; RECORD_SIZE];
-        buf[0..4].copy_from_slice(&SECTOR_MAGIC);
-        buf[4..8].copy_from_slice(&self.erase_count.to_le_bytes());
-        buf[8] = self.next_slot;
-        buf
-    }
-
-    pub fn from_bytes(buf: &RecordBytes) -> Result<Self, SectorDecodeError> {
-        if buf[0..4] != SECTOR_MAGIC {
-            return Err(SectorDecodeError::BadMagic);
-        }
-        Ok(Self {
-            erase_count: u32::from_le_bytes(buf[4..8].try_into().unwrap()),
-            next_slot: buf[8],
-        })
-    }
-
-    pub fn is_full(&self) -> bool {
-        self.next_slot as usize >= MAX_RECORDS_PER_SECTOR || self.next_slot == SECTOR_FULL
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum SectorDecodeError {
-    BadMagic,
-}
-
-// ---------------------------------------------------------------------------
-// CrashContext global — updated by the emulator loop (core 0), read-only by
-// the fault handler.  All fields are AtomicU32 for lock-free cross-core access.
-//
-// `valid` is written LAST with Release ordering; all other stores use Relaxed.
-// The fault handler reads `valid` with Acquire, then reads other fields with
-// Relaxed.  If `valid == 0` the fault handler leaves GB fields zeroed.
-// ---------------------------------------------------------------------------
-
-use core::sync::atomic::{AtomicU32, Ordering};
-
-#[cfg(target_arch = "arm")]
-pub struct CrashContext {
-    /// Non-zero when a full snapshot has been committed at least once.
-    valid: AtomicU32,
-    /// First 4 bytes of the ROM SHA-256 hash packed as LE u32.
-    rom_id_prefix: AtomicU32,
-    /// [rom_bank:16 | ram_bank:8 | _:8]
-    rom_bank_info: AtomicU32,
-    /// [gb_a:8 | gb_f:8 | gb_b:8 | gb_c:8]
-    gb_af_bc: AtomicU32,
-    /// [gb_d:8 | gb_e:8 | gb_h:8 | gb_l:8]
-    gb_de_hl: AtomicU32,
-    /// [gb_sp:16 | gb_pc:16]
-    gb_sp_pc: AtomicU32,
-    /// Lower 32 bits of the u64 cycle counter.
-    gb_cycle_lo: AtomicU32,
-    /// [ppu_ly:8 | ppu_stat:8 | _:16]
-    ppu_ly_stat: AtomicU32,
-}
-
-#[cfg(target_arch = "arm")]
-impl CrashContext {
-    pub const fn new() -> Self {
-        Self {
-            valid: AtomicU32::new(0),
-            rom_id_prefix: AtomicU32::new(0),
-            rom_bank_info: AtomicU32::new(0),
-            gb_af_bc: AtomicU32::new(0),
-            gb_de_hl: AtomicU32::new(0),
-            gb_sp_pc: AtomicU32::new(0),
-            gb_cycle_lo: AtomicU32::new(0),
-            ppu_ly_stat: AtomicU32::new(0),
-        }
-    }
-
-    /// Called once per frame from the emulator tick loop (core 0).
-    pub fn update(
-        &self,
-        rom_id_prefix: [u8; 4],
-        rom_bank: u16,
-        ram_bank: u8,
-        gb_a: u8,
-        gb_f: u8,
-        gb_b: u8,
-        gb_c: u8,
-        gb_d: u8,
-        gb_e: u8,
-        gb_h: u8,
-        gb_l: u8,
-        gb_sp: u16,
-        gb_pc: u16,
-        gb_cycle_lo: u32,
-        ppu_ly: u8,
-        ppu_stat: u8,
-    ) {
-        // Invalidate while updating so the fault handler can't read a torn snapshot.
-        self.valid.store(0, Ordering::Release);
-
-        self.rom_id_prefix
-            .store(u32::from_le_bytes(rom_id_prefix), Ordering::Relaxed);
-        self.rom_bank_info.store(
-            (rom_bank as u32) | ((ram_bank as u32) << 16),
-            Ordering::Relaxed,
-        );
-        self.gb_af_bc.store(
-            (gb_a as u32) | ((gb_f as u32) << 8) | ((gb_b as u32) << 16) | ((gb_c as u32) << 24),
-            Ordering::Relaxed,
-        );
-        self.gb_de_hl.store(
-            (gb_d as u32) | ((gb_e as u32) << 8) | ((gb_h as u32) << 16) | ((gb_l as u32) << 24),
-            Ordering::Relaxed,
-        );
-        self.gb_sp_pc
-            .store((gb_sp as u32) | ((gb_pc as u32) << 16), Ordering::Relaxed);
-        self.gb_cycle_lo.store(gb_cycle_lo, Ordering::Relaxed);
-        self.ppu_ly_stat.store(
-            (ppu_ly as u32) | ((ppu_stat as u32) << 8),
-            Ordering::Relaxed,
-        );
-
-        // Publish: Release fence ensures all stores above are visible to any
-        // subsequent Acquire load of `valid`.
-        self.valid.store(1, Ordering::Release);
-    }
-
-    /// Read the snapshot atomically.  Returns `None` if no update has been
-    /// committed (e.g. crash happened before the first game frame).
-    pub fn snapshot(&self) -> Option<CrashContextSnapshot> {
-        if self.valid.load(Ordering::Acquire) == 0 {
-            return None;
-        }
-        let rom_id_raw = self.rom_id_prefix.load(Ordering::Relaxed);
-        let bank_info = self.rom_bank_info.load(Ordering::Relaxed);
-        let af_bc = self.gb_af_bc.load(Ordering::Relaxed);
-        let de_hl = self.gb_de_hl.load(Ordering::Relaxed);
-        let sp_pc = self.gb_sp_pc.load(Ordering::Relaxed);
-        let cycle_lo = self.gb_cycle_lo.load(Ordering::Relaxed);
-        let ly_stat = self.ppu_ly_stat.load(Ordering::Relaxed);
-
-        Some(CrashContextSnapshot {
-            rom_id_prefix: rom_id_raw.to_le_bytes(),
-            rom_bank: (bank_info & 0xFFFF) as u16,
-            ram_bank: ((bank_info >> 16) & 0xFF) as u8,
-            gb_a: (af_bc & 0xFF) as u8,
-            gb_f: ((af_bc >> 8) & 0xFF) as u8,
-            gb_b: ((af_bc >> 16) & 0xFF) as u8,
-            gb_c: ((af_bc >> 24) & 0xFF) as u8,
-            gb_d: (de_hl & 0xFF) as u8,
-            gb_e: ((de_hl >> 8) & 0xFF) as u8,
-            gb_h: ((de_hl >> 16) & 0xFF) as u8,
-            gb_l: ((de_hl >> 24) & 0xFF) as u8,
-            gb_sp: (sp_pc & 0xFFFF) as u16,
-            gb_pc: ((sp_pc >> 16) & 0xFFFF) as u16,
-            gb_cycle_lo: cycle_lo,
-            ppu_ly: (ly_stat & 0xFF) as u8,
-            ppu_stat: ((ly_stat >> 8) & 0xFF) as u8,
-        })
-    }
-}
-
-#[cfg(target_arch = "arm")]
-unsafe impl Sync for CrashContext {}
-
-/// Snapshot of the emulator state captured by the fault handler.
-#[derive(Debug, Clone, Copy)]
-pub struct CrashContextSnapshot {
-    pub rom_id_prefix: [u8; 4],
-    pub rom_bank: u16,
-    pub ram_bank: u8,
-    pub gb_a: u8,
-    pub gb_f: u8,
-    pub gb_b: u8,
-    pub gb_c: u8,
-    pub gb_d: u8,
-    pub gb_e: u8,
-    pub gb_h: u8,
-    pub gb_l: u8,
-    pub gb_sp: u16,
-    pub gb_pc: u16,
-    pub gb_cycle_lo: u32,
-    pub ppu_ly: u8,
-    pub ppu_stat: u8,
-}
-
-/// The single global crash context, updated by the emulator loop.
-#[cfg(target_arch = "arm")]
-pub static CRASH_CONTEXT: CrashContext = CrashContext::new();
-
-/// Durable diagnostic captured immediately before `report_transport_smash`
-/// panics. The panic handler consumes this and stores the values in the ARM
-/// fields of a `TransportSmash` record:
-/// - `arm_pc` = `Core1Transport` base
-/// - `arm_lr` = corrupted `command_tx` pointer
-/// - `arm_cfsr` = corrupted `audio_rx` pointer
-/// - `arm_hfsr` = corrupted `shared` pointer
-/// - `arm_fault_addr` = first duplicate triplet found in SRAM, or 0
-pub struct TransportSmashDiag {
-    active: AtomicU32,
-    base: AtomicU32,
-    cmd: AtomicU32,
-    aud: AtomicU32,
-    shr: AtomicU32,
-    source_triplet: AtomicU32,
-}
-
-impl TransportSmashDiag {
-    pub const fn new() -> Self {
-        Self {
-            active: AtomicU32::new(0),
-            base: AtomicU32::new(0),
-            cmd: AtomicU32::new(0),
-            aud: AtomicU32::new(0),
-            shr: AtomicU32::new(0),
-            source_triplet: AtomicU32::new(0),
-        }
-    }
-
-    pub fn record(&self, base: usize, cmd: usize, aud: usize, shr: usize, source_triplet: usize) {
-        self.active.store(0, Ordering::Release);
-        self.base.store(base as u32, Ordering::Relaxed);
-        self.cmd.store(cmd as u32, Ordering::Relaxed);
-        self.aud.store(aud as u32, Ordering::Relaxed);
-        self.shr.store(shr as u32, Ordering::Relaxed);
-        self.source_triplet
-            .store(source_triplet as u32, Ordering::Relaxed);
-        self.active.store(1, Ordering::Release);
-    }
-
-    pub fn take(&self) -> Option<TransportSmashSnapshot> {
-        if self.active.swap(0, Ordering::AcqRel) == 0 {
-            return None;
-        }
-        Some(TransportSmashSnapshot {
-            base: self.base.load(Ordering::Relaxed),
-            cmd: self.cmd.load(Ordering::Relaxed),
-            aud: self.aud.load(Ordering::Relaxed),
-            shr: self.shr.load(Ordering::Relaxed),
-            source_triplet: self.source_triplet.load(Ordering::Relaxed),
-        })
-    }
-}
-
-unsafe impl Sync for TransportSmashDiag {}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct TransportSmashSnapshot {
-    pub base: u32,
-    pub cmd: u32,
-    pub aud: u32,
-    pub shr: u32,
-    pub source_triplet: u32,
-}
-
-pub static TRANSPORT_SMASH_DIAG: TransportSmashDiag = TransportSmashDiag::new();
-
-// ---------------------------------------------------------------------------
-// CRC32 (IEEE 802.3 polynomial, table-free).
+// CRC32 (IEEE 802.3 polynomial, table-free) — shared by all sub-modules.
 // ---------------------------------------------------------------------------
 
 pub fn crc32(data: &[u8]) -> u32 {
@@ -678,102 +69,25 @@ pub fn crc32(data: &[u8]) -> u32 {
 }
 
 // ---------------------------------------------------------------------------
-// Scratch register helper — maps WATCHDOG+POWMAN scratch to CrashRecord.
-// Used by storage.rs at boot time.
+// Re-exports: keep all existing public paths working unchanged.
 // ---------------------------------------------------------------------------
 
-/// Sixteen u32 values read from hardware scratch registers.
-///
-/// Index 0-7  → WATCHDOG.scratch0-7
-/// Index 8-15 → POWMAN.scratch(0)-scratch(7)
-pub struct ScratchRegs(pub [u32; 16]);
+pub use record::{
+    flags, CrashDecodeError, CrashKind, CrashRecord, RecordBytes, MAX_RECORDS_PER_SECTOR,
+    SECTOR_HEADER_SIZE,
+};
 
-impl ScratchRegs {
-    pub fn is_crash(&self) -> bool {
-        self.0[0] == CRASH_MAGIC
-    }
+pub use sector::{find_next_empty_slot, SectorDecodeError, SectorHeader, SECTOR_FULL};
 
-    /// Build a [`CrashRecord`] from the scratch blob + build-time constants.
-    /// `slot_seq` is supplied by the storage layer.
-    pub fn to_crash_record(&self, slot_seq: u8) -> CrashRecord {
-        let wd = &self.0;
-        // WATCHDOG scratch layout:
-        //   [0] = CRASH_MAGIC
-        //   [1] = arm_pc
-        //   [2] = arm_lr
-        //   [3] = arm_cfsr
-        //   [4] = arm_hfsr
-        //   [5] = arm_fault_addr
-        //   [6] = [crash_kind:8 | flags:8 | ram_bank:8 | _:8]
-        //   [7] = [rom_bank:16 | gb_pc:16]
-        let crash_kind = (wd[6] & 0xFF) as u8;
-        let flags_byte = ((wd[6] >> 8) & 0xFF) as u8;
-        let ram_bank = ((wd[6] >> 16) & 0xFF) as u8;
-        let rom_bank = (wd[7] >> 16) as u16;
-        let gb_pc = (wd[7] & 0xFFFF) as u16;
+pub use context::{
+    ScratchRegs, TransportSmashDiag, TransportSmashSnapshot, TRANSPORT_SMASH_DIAG,
+};
 
-        // POWMAN scratch layout (offset by 8):
-        //   [8]  = rom_id_prefix
-        //   [9]  = [gb_a:8 | gb_f:8 | gb_b:8 | gb_c:8]
-        //   [10] = [gb_d:8 | gb_e:8 | gb_h:8 | gb_l:8]
-        //   [11] = [gb_sp:16 | ppu_ly:8 | ppu_stat:8]
-        //   [12] = gb_cycle_lo
-        //   [13] = [stack_headroom:16 | panic_line_or_r1:16]
-        //   [14] = panic_file[0..4]
-        //   [15] = panic_file[4..8]
-        let rom_id_prefix = wd[8].to_le_bytes();
-        let af_bc = wd[9];
-        let de_hl = wd[10];
-        let sp_ly = wd[11];
-        let gb_cycle_lo = wd[12];
-        let panic_line = (wd[13] & 0xFFFF) as u16;
-        let stack_headroom = (wd[13] >> 16) as u16;
-
-        let mut panic_loc = [0u8; 12];
-        panic_loc[0..4].copy_from_slice(&wd[14].to_le_bytes());
-        panic_loc[4..8].copy_from_slice(&wd[15].to_le_bytes());
-
-        CrashRecord {
-            schema_ver: 1,
-            crash_kind,
-            flags: flags_byte,
-            slot_seq,
-            fw_version: FW_VERSION,
-            git_hash: GIT_HASH_U32,
-            arm_pc: wd[1],
-            arm_lr: wd[2],
-            arm_cfsr: wd[3],
-            arm_hfsr: wd[4],
-            arm_fault_addr: wd[5],
-            rom_id_prefix,
-            rom_bank,
-            ram_bank,
-            gb_a: (af_bc & 0xFF) as u8,
-            gb_f: ((af_bc >> 8) & 0xFF) as u8,
-            gb_b: ((af_bc >> 16) & 0xFF) as u8,
-            gb_c: ((af_bc >> 24) & 0xFF) as u8,
-            gb_d: (de_hl & 0xFF) as u8,
-            gb_e: ((de_hl >> 8) & 0xFF) as u8,
-            gb_h: ((de_hl >> 16) & 0xFF) as u8,
-            gb_l: ((de_hl >> 24) & 0xFF) as u8,
-            gb_sp: (sp_ly & 0xFFFF) as u16,
-            gb_pc,
-            gb_cycle_lo,
-            ppu_ly: ((sp_ly >> 16) & 0xFF) as u8,
-            ppu_lcdc: 0, // not captured in scratch
-            ppu_stat: ((sp_ly >> 24) & 0xFF) as u8,
-            panic_loc,
-            panic_line,
-            stack_headroom,
-            // DMA fields are injected after this call in check_and_commit,
-            // once the .uninit DMA_CRASH_SNAPSHOT has been read.
-            ..Default::default()
-        }
-    }
-}
+#[cfg(target_arch = "arm")]
+pub use context::{CrashContext, CrashContextSnapshot, CRASH_CONTEXT};
 
 // ---------------------------------------------------------------------------
-// Tests
+// Tests — kept in mod.rs because they test types across multiple sub-modules.
 // ---------------------------------------------------------------------------
 
 #[cfg(test)]
